Rename transport.rs to client.rs
See previous commit.
I did as 2 commits to ensure git recognized that
`transport.rs` is renamed to `client.rs`
