+++ libs/gl-client/src/lsps/mod.rs
@@ -1,5 +1,5 @@
 pub mod json_rpc;
 pub mod json_rpc_erased;
 pub mod message;
-pub mod transport;
+pub mod client;
 pub mod error;
